agents/gemini-enterprise/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-aiplatform
